test_simple: make `test_ssh2_dh_validate()` portable
Omit direct backend-specific API use. Drop one test using negative
input, that's impossible with the internal crypto API. Using
backend-specific code in tests is highly undesired. It's also
discouraged to special-case any, in this case WinCNG had to be because
it does not implement the DH validation function.

Also:
- mbedtls: fix an internal API type difference with a cast.
- unfold a line.

Follow-up to fcf830788d0c8dcd6bed7dcd532506aabf196523 #2239

Closes #2595

+++ src/mbedtls.h
@@ -218,7 +218,7 @@ struct ssh2_mbed_cipher_ctx {
  */
 
 #define ssh2_bn                        mbedtls_mpi
-#define ssh2_bn_set_word(bn, word)     mbedtls_mpi_lset(bn, word)
+#define ssh2_bn_set_word(bn, word)     mbedtls_mpi_lset(bn, (long long)(word))
 #define ssh2_bn_to_bin(bn, bin) \
     mbedtls_mpi_write_binary(bn, bin, mbedtls_mpi_size(bn))
 #define ssh2_bn_bytes(bn)              mbedtls_mpi_size(bn)
